chore: Move knex-specific database adapter code out of EntityDatabaseAdapter
