+++ packages/entity/src/utils/__testfixtures__/StubDatabaseAdapterProvider.ts
@@ -1,7 +1,9 @@
 import { EntityConfiguration } from '../../EntityConfiguration';
 import { EntityDatabaseAdapter } from '../../EntityDatabaseAdapter';
+import { EntityKnexDatabaseAdapter } from '../../EntityKnexDatabaseAdapter';
 import { IEntityDatabaseAdapterProvider } from '../../IEntityDatabaseAdapterProvider';
 import { StubDatabaseAdapter } from '../__testfixtures__/StubDatabaseAdapter';
+import { StubKnexDatabaseAdapter } from '../__testfixtures__/StubKnexDatabaseAdapter';
 
 export class StubDatabaseAdapterProvider implements IEntityDatabaseAdapterProvider {
   private readonly objectCollection = new Map();
@@ -10,5 +12,11 @@ export class StubDatabaseAdapterProvider implements IEntityDatabaseAdapterProvid
     entityConfiguration: EntityConfiguration<TFields, TIDField>,
   ): EntityDatabaseAdapter<TFields, TIDField> {
     return new StubDatabaseAdapter(entityConfiguration, this.objectCollection);
+  }
+
+  getKnexDatabaseAdapter<TFields extends Record<string, any>, TIDField extends keyof TFields>(
+    entityConfiguration: EntityConfiguration<TFields, TIDField>,
+  ): EntityKnexDatabaseAdapter<TFields, TIDField> {
+    return new StubKnexDatabaseAdapter(entityConfiguration, new Map(), this.objectCollection);
   }
 }
